PointCloud. Highlighting
